finish feature order in seller, add screen statistic

+++ app/src/main/java/com/example/testbase/model/Cart.kt
@@ -6,6 +6,6 @@ data class Cart(
     var id: Int? = null,
     var createAt: Long = 0,
     var totalPrice: Long = 0,
-    var totalQuantity: Long = 0,
+    var totalQuantity: Int = 0,
     var user: User = User()
 ): Serializable
